Use u64 to decode metadata value

+++ refine-edh/src/phb/data/metadata.rs
@@ -3,5 +3,5 @@ use serde::Deserialize;
 #[derive(Deserialize)]
 pub(in crate::phb) struct PMetadata {
     pub(in crate::phb) field_name: String,
-    pub(in crate::phb) field_value: u32,
+    pub(in crate::phb) field_value: u64,
 }
